fix(#1920): light package __init__ + E2E visual-grounding test — resolve re-review blockers
Blocker 1 — bridge transitively loaded the heavy pipeline: importing
tools.video_watch.constants executes the package __init__, which eagerly
imported tools.link_analysis + grok (httpx, Pillow path) at bridge import
time; the grep-based R8 row passed vacuously. Restructure:
- tools/video_watch/pipeline.py — the heavy pull pipeline (watch_video,
  detect_source, VideoWatchError, yt-dlp/ffmpeg/Pillow/Whisper)
- tools/video_watch/reaper.py — stdlib-only reap_stale_frame_dirs, importable
  by cli.py and agent/session_health.py without the pipeline
- tools/video_watch/__init__.py — LIGHT: stdlib + constants re-export only;
  public callables resolve lazily via PEP 562 __getattr__ (back-compat kept)
- Plan Verification R8 hardened from the vacuous grep to an import-chain
  assertion: tests/test_import_discipline.py runs fresh-interpreter probes
  (constants import, reaper import, bridge.enrichment import) asserting none
  of pipeline/grok/link_analysis/PIL/httpx load

Blocker 2 — E2E visual-grounding criterion had no committed evidence:
- tests/test_e2e_visual_grounding.py synthesizes a silent 3-slide deck
  (ffmpeg lavfi, hard cuts, silent audio track) and runs the REAL
  ffprobe/ffmpeg/Pillow pipeline with only the two network edges patched;
  asserts transcript-only fails while emitted frames are multiple,
  persistent, pairwise visually distinct, and temporally ordered
- Plan Success Criterion updated to cite the committed evidence

Tech debt — forward cross-links added from youtube-transcription.md and
markitdown-ingestion.md to the watch tier; feature doc file map updated to
the new module structure

39 feature tests pass (was 33); ruff check/format clean
